refactor: structure files into folders for clearer separation
