+++ src/main/java/nl/hartwigmedicalfoundation/rps/model/GameResult.java
@@ -1,4 +1,4 @@
-package nl.hartwigmedicalfoundation.rps;
+package nl.hartwigmedicalfoundation.rps.model;
 
 public enum GameResult {
     WIN_PLAYER_1, WIN_PLAYER_2, DRAW
