+++ src/main/java/nl/hartwigmedicalfoundation/rps/model/Move.java
@@ -1,4 +1,4 @@
-package nl.hartwigmedicalfoundation.rps;
+package nl.hartwigmedicalfoundation.rps.model;
 
 public enum Move {
     ROCK, PAPER, SCISSORS;
